refactor(core/l10n): move getIntlData to core/utils
